Add more content to tests and merge notification handler into callbackhandler
