+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/caseworker/MoreTimeRequestedCallbackHandler.java
@@ -2,26 +2,42 @@ package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.caseworker;
 
 import com.google.common.collect.ImmutableMap;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.cmc.ccd.domain.CCDCase;
+import uk.gov.hmcts.cmc.ccd.domain.CCDCollectionElement;
+import uk.gov.hmcts.cmc.ccd.domain.CCDContactPartyType;
+import uk.gov.hmcts.cmc.ccd.domain.CCDDocument;
+import uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption;
 import uk.gov.hmcts.cmc.ccd.domain.CaseEvent;
+import uk.gov.hmcts.cmc.ccd.domain.GeneralLetterContent;
+import uk.gov.hmcts.cmc.ccd.domain.defendant.CCDRespondent;
+import uk.gov.hmcts.cmc.claimstore.config.properties.notifications.NotificationsProperties;
 import uk.gov.hmcts.cmc.claimstore.events.EventProducer;
-import uk.gov.hmcts.cmc.claimstore.events.response.MoreTimeRequestedCitizenNotificationHandler;
+import uk.gov.hmcts.cmc.claimstore.idam.models.UserDetails;
 import uk.gov.hmcts.cmc.claimstore.rules.MoreTimeRequestRule;
 import uk.gov.hmcts.cmc.claimstore.services.ResponseDeadlineCalculator;
+import uk.gov.hmcts.cmc.claimstore.services.UserService;
+import uk.gov.hmcts.cmc.claimstore.services.ccd.DocAssemblyService;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.Role;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.Callback;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackHandler;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackParams;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackType;
+import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.generalletter.GeneralLetterService;
+import uk.gov.hmcts.cmc.claimstore.services.notifications.NotificationService;
 import uk.gov.hmcts.cmc.claimstore.utils.CaseDetailsConverter;
 import uk.gov.hmcts.cmc.domain.models.Claim;
+import uk.gov.hmcts.cmc.domain.utils.PartyUtils;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse.AboutToStartOrSubmitCallbackResponseBuilder;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
+import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+import uk.gov.hmcts.reform.docassembly.domain.DocAssemblyResponse;
 
+import java.net.URISyntaxException;
 import java.time.LocalDate;
 import java.util.Collections;
 import java.util.HashMap;
@@ -30,6 +46,14 @@ import java.util.Map;
 
 import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.RESPONSE_MORE_TIME;
 import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.CASEWORKER;
+import static uk.gov.hmcts.cmc.claimstore.services.notifications.NotificationReferenceBuilder.MoreTimeRequested.referenceForClaimant;
+import static uk.gov.hmcts.cmc.claimstore.services.notifications.NotificationReferenceBuilder.MoreTimeRequested.referenceForDefendant;
+import static uk.gov.hmcts.cmc.claimstore.services.notifications.content.NotificationTemplateParameters.CLAIMANT_NAME;
+import static uk.gov.hmcts.cmc.claimstore.services.notifications.content.NotificationTemplateParameters.CLAIMANT_TYPE;
+import static uk.gov.hmcts.cmc.claimstore.services.notifications.content.NotificationTemplateParameters.CLAIM_REFERENCE_NUMBER;
+import static uk.gov.hmcts.cmc.claimstore.services.notifications.content.NotificationTemplateParameters.DEFENDANT_NAME;
+import static uk.gov.hmcts.cmc.claimstore.services.notifications.content.NotificationTemplateParameters.FRONTEND_BASE_URL;
+import static uk.gov.hmcts.cmc.claimstore.services.notifications.content.NotificationTemplateParameters.RESPONSE_DEADLINE;
 import static uk.gov.hmcts.cmc.claimstore.utils.Formatting.formatDate;
 
 @Service
@@ -44,11 +68,23 @@ public class MoreTimeRequestedCallbackHandler extends CallbackHandler {
     private final ResponseDeadlineCalculator responseDeadlineCalculator;
     private final MoreTimeRequestRule moreTimeRequestRule;
     private final CaseDetailsConverter caseDetailsConverter;
-    private final MoreTimeRequestedCitizenNotificationHandler moreTimeRequestedCitizenNotificationHandler;
+    private final NotificationService notificationService;
+    private final NotificationsProperties notificationsProperties;
+    private final GeneralLetterService generalLetterService;
+    private final DocAssemblyService docAssemblyService;
+    private final UserService userService;
+    private final String generalLetterTemplateId;
 
     private static final String PREVIEW_SENTENCE = "The response deadline will be %s .";
     private static final String ERROR_MESSAGE =
         "There was a technical problem. Nothing has been sent. You need to try again.";
+    private static final String STANDARD_DEADLINE_TEXT = "You’ve been given an extra 14 days to respond to the"
+            + " claim made against you by %s.\n"
+            + "\n"
+            + "You now need to respond to the claim before 4pm on %s.\n"
+            + "\n"
+            + "If you don’t respond, you could get a County Court Judgment (CCJ). This may make it harder to get "
+            + "credit, such as a mobile phone contract, credit card or mortgage.";
 
     @Autowired
     public MoreTimeRequestedCallbackHandler(
@@ -56,12 +92,22 @@ public class MoreTimeRequestedCallbackHandler extends CallbackHandler {
             ResponseDeadlineCalculator responseDeadlineCalculator,
             MoreTimeRequestRule moreTimeRequestRule,
             CaseDetailsConverter caseDetailsConverter,
-            MoreTimeRequestedCitizenNotificationHandler moreTimeRequestedCitizenNotificationHandler) {
+            NotificationService notificationService,
+            NotificationsProperties notificationsProperties,
+            GeneralLetterService generalLetterService,
+            DocAssemblyService docAssemblyService,
+            UserService userService,
+            @Value("${doc_assembly.generalLetterTemplateId}") String generalLetterTemplateId) {
         this.eventProducer = eventProducer;
         this.responseDeadlineCalculator = responseDeadlineCalculator;
         this.moreTimeRequestRule = moreTimeRequestRule;
         this.caseDetailsConverter = caseDetailsConverter;
-        this.moreTimeRequestedCitizenNotificationHandler = moreTimeRequestedCitizenNotificationHandler;
+        this.generalLetterService = generalLetterService;
+        this.notificationService = notificationService;
+        this.notificationsProperties = notificationsProperties;
+        this.docAssemblyService = docAssemblyService;
+        this.userService = userService;
+        this.generalLetterTemplateId = generalLetterTemplateId;
     }
 
     @Override
@@ -83,22 +129,123 @@ public class MoreTimeRequestedCallbackHandler extends CallbackHandler {
     }
 
     public CallbackResponse sendNotifications(CallbackParams callbackParams) {
-        Claim claim = convertCallbackToClaim(callbackParams.getRequest());
+        CallbackRequest callbackRequest = callbackParams.getRequest();
+        CaseDetails caseDetails = callbackRequest.getCaseDetails();
+        Claim claim = caseDetailsConverter.extractClaim(caseDetails);
+        LocalDate responseDeadline = responseDeadlineCalculator.calculatePostponedResponseDeadline(claim.getIssuedOn());
+        CallbackResponse response = AboutToStartOrSubmitCallbackResponse.builder().build();
 
-        eventProducer.createMoreTimeForResponseRequestedEvent(
-                claim,
-                claim.getResponseDeadline(),
-                claim.getClaimData().getDefendant().getEmail().orElse(null)
-        );
         try {
-            return moreTimeRequestedCitizenNotificationHandler.sendNotifications(callbackParams);
+            eventProducer.createMoreTimeForResponseRequestedEvent(
+                    claim,
+                    claim.getResponseDeadline(),
+                    claim.getClaimData().getDefendant().getEmail().orElse(null)
+            );
+
+            if (claim.getDefendantEmail() != null && claim.getDefendantId() != null) {
+                sendEmailNotificationToDefendant(claim, responseDeadline);
+            } else {
+                response = createAndPrintLetter(callbackParams);
+            }
+            sendNotificationToClaimant(claim, responseDeadline);
+
+            return response;
         } catch (Exception e) {
             return AboutToStartOrSubmitCallbackResponse
-               .builder()
-                .errors(Collections.singletonList(ERROR_MESSAGE))
-                .build();
+                    .builder()
+                    .errors(Collections.singletonList(ERROR_MESSAGE))
+                    .build();
         }
+    }
 
+    private void sendEmailNotificationToDefendant(Claim claim, LocalDate deadline) {
+        notificationService.sendMail(
+                claim.getDefendantEmail(),
+                notificationsProperties.getTemplates().getEmail().getDefendantMoreTimeRequested(),
+                prepareNotificationParameters(claim,deadline),
+                referenceForDefendant(claim.getReferenceNumber())
+        );
+    }
+
+    private void sendNotificationToClaimant(Claim claim, LocalDate deadline) {
+        notificationService.sendMail(
+                claim.getSubmitterEmail(),
+                notificationsProperties.getTemplates().getEmail().getClaimantMoreTimeRequested(),
+                prepareNotificationParameters(claim, deadline),
+                referenceForClaimant(claim.getReferenceNumber())
+        );
+    }
+
+    private CallbackResponse createAndPrintLetter(CallbackParams callbackParams) throws URISyntaxException {
+
+        String authorisation = callbackParams.getParams().get(CallbackParams.Params.BEARER_TOKEN).toString();
+        UserDetails userDetails = userService.getUserDetails(authorisation);
+        String caseworkerName = userDetails.getFullName();
+        CCDCase ccdCase = caseDetailsConverter.extractCCDCase(callbackParams.getRequest().getCaseDetails());
+        Claim claim = caseDetailsConverter.extractClaim(callbackParams.getRequest().getCaseDetails());
+        LocalDate responseDeadline = responseDeadlineCalculator.calculatePostponedResponseDeadline(claim.getIssuedOn());
+
+        GeneralLetterContent generalLetterContent = GeneralLetterContent.builder()
+                .caseworkerName(caseworkerName)
+                .letterContent(String.format(STANDARD_DEADLINE_TEXT, claim.getClaimData().getClaimant().getName(),
+                        formatDate(responseDeadline)))
+                .issueLetterContact(CCDContactPartyType.DEFENDANT)
+                .build();
+
+        CCDCase updatedCCDCase = ccdCase.toBuilder()
+                .generalLetterContent(generalLetterContent)
+                .build();
+
+        CCDRespondent respondent = ccdCase.getRespondents().get(0).getValue().toBuilder()
+                .responseMoreTimeNeededOption(CCDYesNoOption.YES)
+                .responseDeadline(responseDeadline)
+                .build();
+
+        DocAssemblyResponse docAssemblyResponse = docAssemblyService
+                .createGeneralLetter(updatedCCDCase, authorisation, generalLetterTemplateId);
+        CCDDocument ccdDocument = CCDDocument.builder()
+                .documentUrl(docAssemblyResponse.getRenditionOutputLocation())
+                .documentBinaryUrl(docAssemblyResponse.getRenditionOutputLocation() + "/binary")
+                .documentFileName("")
+                .build();
+        updatedCCDCase.setDraftLetterDoc(ccdDocument);
+        generalLetterService.printLetter(authorisation, updatedCCDCase.getDraftLetterDoc(), claim);
+        CCDCase updatedCase = ccdCase.toBuilder()
+                .caseDocuments(generalLetterService
+                        .updateCaseDocumentsWithGeneralLetter(
+                                updatedCCDCase,
+                                updatedCCDCase.getDraftLetterDoc(),
+                                String.format("%s-response-deadline-extended.pdf",
+                                        claim.getReferenceNumber())))
+                .generalLetterContent(null)
+                .draftLetterDoc(null)
+                .respondents(List.of(CCDCollectionElement.<CCDRespondent>builder()
+                        .value(respondent)
+                        .build()))
+                .build();
+
+        return AboutToStartOrSubmitCallbackResponse
+                .builder()
+                .data(caseDetailsConverter.convertToMap(updatedCase))
+                .build();
+//        } catch (Exception e) {
+//            e.getStackTrace();
+//            return AboutToStartOrSubmitCallbackResponse
+//                .builder()
+//                .errors(Collections.singletonList(ERROR_MESSAGE))
+//                .build();
+//        }
+    }
+
+    private Map<String, String> prepareNotificationParameters(Claim claim, LocalDate deadline) {
+        Map<String, String> parameters = new HashMap<>();
+        parameters.put(CLAIM_REFERENCE_NUMBER, claim.getReferenceNumber());
+        parameters.put(CLAIMANT_TYPE, PartyUtils.getType(claim.getClaimData().getClaimant()));
+        parameters.put(CLAIMANT_NAME, claim.getClaimData().getClaimant().getName());
+        parameters.put(DEFENDANT_NAME, claim.getClaimData().getDefendant().getName());
+        parameters.put(RESPONSE_DEADLINE, formatDate(deadline));
+        parameters.put(FRONTEND_BASE_URL, notificationsProperties.getFrontendBaseUrl());
+        return parameters;
     }
 
     private AboutToStartOrSubmitCallbackResponse requestMoreTimeViaCaseworker(CallbackParams callbackParams) {
