+++ src/test/java/uk/gov/hmcts/cmc/claimstore/services/ClaimServiceTest.java
@@ -129,7 +129,7 @@ public class ClaimServiceTest {
             userService,
             issueDateCalculator,
             responseDeadlineCalculator,
-            new MoreTimeRequestRule(new ClaimDeadlineService()),
+            new MoreTimeRequestRule(),
             eventProducer,
             appInsights,
             new PaidInFullRule(),
@@ -188,7 +188,7 @@ public class ClaimServiceTest {
             userService,
             issueDateCalculator,
             responseDeadlineCalculator,
-            new MoreTimeRequestRule(new ClaimDeadlineService()),
+            new MoreTimeRequestRule(),
             eventProducer,
             appInsights,
             new PaidInFullRule(),
